implent timer as service
It seems that using a service to check for timer expiration is
the way to go.
No job scheduling yet, using Handler + Runnable works and it has
the advantage that I don't need to implement a messenger.
Although, possibly, the ringtone could be handled by a job scheduler.
Accelerometer checking should also be a scheduler.