update sleep time handling
Calculating sleep time is proving more complicated that expected.
I didn't cover the case where the "expected pattern" (wake up in
the morning go to sleep before midnight) is inverted, resulting
in a sleeping time "preceding" a wake-up time.
It's not difficult to solve, per se, and this patch should handle
it, but I still have to think of a solution simple enough to be
happy with.