try to fix wake-up issues
Still the services seem not to run as I expect them to do; they seem to
work as intended when trying them in the day, but they fail when the
phone is left alone for a long period of time (like night sleeping).
I don't know the cause yet and I have no viable hypothesis so far, so
trying to poke around and see what happens (the downside is that it
takes a long time to verify).