add wake-lock
There were uncertainties in the service executions due to the phone
going to sleep. I have now added a wake-lock on the CPU to see if it
improves the application matching the timers.
Another solution, suggested by the Android documentation, is to use
the job scheduler, but I am not sure I want to to do that as it will
leave me *more* dependent on the Android OS.