move orientation code
Moved the code detecting orientation change into the service detecting
the movements of the phone.
Not sure it's a good idea overall (or a better idea than having its own
service), but it's driven by the desire to split the services in two
categories: a "polling" one (check changes, reset on changes over a
threshold, sleep for a bit) and a "triggered" one (reset on any event
triggered).