add sleeping time call before notification
Small thing: I seem not able to get the sleeping time to kick in
before the time when the expiration is going to be past it.
I am not sure the reason. I must be missing something in the code,
or, and the currect code should fix it, it was just the notification
not being correct or up to date.
We'll see.