update sleeping date calculations
I think this may (if my code is sound and I didn't miss something)
fix the issue with the service not going to sleep on occasion.
The sleeping and waking up dates were not being updated correctly after
the first day (that's one reason why it was difficult to find out).