refactor sleep interval code
Still not happy about the whole way to detect and calculate sleeping
time.
I am trying to separate and simplify the calculations and to reduce
the calls to _isSleepTime_ method.
After keeping an eye on this commit changes, I will re-evaluate
the situation (I know already some part I want to change, but they
are minor and they should not change the underlying logic).