add toggle buttons
I found that in everyday use there are situation where the system
doesn't seem to work as I would expect.
I thought to add some buttons to disable specific mechanisms to
reset the timer, to find out which one, if any, is responsible
for the behaviour.
It proved a bit more complicated that I was expecting and, now,
because I didn't want to redesign it all, the code is even more
cludgy.